Fixed sorting issues
deleteFile deletes file provided (now gets the full path).
maxThis(String thisUser, Table thisTable) takes a user and a file with
tweets for that users and finds the max ID.

I guess it doesn’t have to get a user - I should write that out and
replace it with anything.